Removed all the TODOs and cruft and bad code. Everything now working for RETR for Cyberduck and command line ftp.

+++ lftpd_string.c
@@ -5,10 +5,10 @@
 
 char* lftpd_string_trim(char* s) {
 	char* p = s;
-	for (int i = 0, len = strlen(s); i < len && isspace(s[i]); i++) {
+	for (int i = 0, len = strlen(s); i < len && isspace((int) s[i]); i++) {
 		p++;
 	}
-	for (int i = strlen(p); i >= 0 && isspace(p[i]); i--) {
+	for (int i = strlen(p); i >= 0 && isspace((int) p[i]); i--) {
 		p[i] = '\0';
 	}
 	return p;
